fix: do not delete special paragraphs

+++ src/services/cms/paragraphs/operations.ts
@@ -3,6 +3,7 @@ import { defineSubOperation } from '@/services/serviceOperation'
 import { cmsParagraphSchemas } from './schemas'
 import { z } from 'zod'
 import { SpecialCmsParagraph } from '@prisma/client'
+import { ServerError } from '@/services/error'
 
 const create = defineSubOperation({
     dataSchema: () => cmsParagraphSchemas.create,
@@ -15,17 +16,17 @@ export const cmsParagraphOperations = {
     destroy: defineSubOperation({
         paramsSchema: () => z.object({ id: z.number() }),
         operation: () => async ({ params, prisma }) => {
+            const paragraph = await prisma.cmsParagraph.findUniqueOrThrow({ where: { id: params.id } })
+            if (paragraph.special) {
+                throw new ServerError('BAD PARAMETERS', 'Special paragraphs cannot be deleted')
+            }
             await prisma.cmsParagraph.delete({ where: { id: params.id } })
         }
     }),
 
     readSpecial: defineSubOperation({
-        paramsSchema: (specialsAllowedToBeReadFromImplementation: SpecialCmsParagraph[]) => z.object({
-            special: z.nativeEnum(SpecialCmsParagraph).refine(
-                val => specialsAllowedToBeReadFromImplementation.includes(val), {
-                    message: 'Special paragraph cannot be accessed through this implementation'
-                }
-            )
+        paramsSchema: () => z.object({
+            special: z.nativeEnum(SpecialCmsParagraph)
         }),
         operation: () => async ({ params, prisma }) => {
             const paragraph = await prisma.cmsParagraph.findUnique({ where: { special: params.special } })
